Unidad4/API/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
easyocr>=1.7.0
pillow>=9.5.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
gunicorn==21.2.0
